Improved Laplace solver
Refined selection of collocation points in the boundary and in the bulk to give the network a better representation of the solution we are looking for. Also selected a better lambda weight so that the boundary and bulk contributions are more balanced. 

The implementation of the optimal lambda is pending.